[EPROD-552] Migrate from ethers-core to alloy

+++ src/evm-canister-client/src/client.rs
@@ -221,7 +221,7 @@ impl<C: CanisterClient> EvmCanisterClient<C> {
     ///
     /// # Returns
     /// The latest block number
-    pub async fn eth_block_number(&self) -> CanisterClientResult<usize> {
+    pub async fn eth_block_number(&self) -> CanisterClientResult<u64> {
         self.client.query("eth_block_number", ()).await
     }
